Rename corresponding file
